+++ src/cmd/cmd_wasm_script.ml
@@ -16,6 +16,11 @@ let cmd_symbolic ~files ~no_exhaustion =
   let exec = Script_symbolic.exec ~no_exhaustion in
   list_iter (run_file exec) files
 
-let cmd_abstract ~files ~no_exhaustion =
+let cmd_abstract ~files ~no_exhaustion ~debug_trace =
+  if Option.is_some debug_trace then Abstract_trace.enable ();
   let exec = Script_abstract.exec ~no_exhaustion in
-  list_iter (run_file exec) files
+  let result = list_iter (run_file exec) files in
+  ( match debug_trace with
+  | None -> ()
+  | Some path -> Abstract_trace.write_json path );
+  result